Extract Prompt Studio canvas forwarding helpers
